# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Philosopher Quote Express",
    "description": "A Philosopher Quote REST API that provides a database of philosopher quotes.",
    "contact": {
      "name": "kdvblended@gmail.com"
    }
  },
  "host": "none",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/authentication/authenticate": {
      "post": {
        "description": "Authenticates the login by email and password.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "A JWT token"
            },
          "400": {
            "description": "Unauthorized!"
          }
          }
        }
      }
  },
...
}